+++ src/main/kotlin/io/github/t45k/part/mining/git/GitCommand.kt
@@ -1,0 +1,7 @@
+package io.github.t45k.part.mining.git
+
+import java.nio.file.Path
+
+abstract class GitCommand<I, O>(protected val projectRootPath: Path, protected val filePath: Path) {
+    abstract fun execute(input: I): O
+}
